delete & upload files and conver pdf to img

+++ user/views.py
@@ -12,16 +12,15 @@ from .forms import RegisterForm,LoginForm,ResetByUsernameForm
 import datetime
 import uuid
 from django.core.validators import validate_email
-import os
 import PyPDF2
 from django.views.decorators.csrf import ensure_csrf_cookie, csrf_exempt
 import json
 from django.db.models import Q
-
+import os, shutil
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-# 登录
+# 登录,DONE
 def login(request):
 	# print(request.POST)
 	if request.session.get('log_in') != None:
@@ -53,7 +52,7 @@ def login(request):
 		return HttpResponse(t.render(c,request))
 
 
-# 注册
+# 注册,DONE
 def register(request):
 	print('register')
 	if request.method =='GET':
@@ -85,14 +84,14 @@ def register(request):
 			c ={'messgage' :'请重新填写资料','form':form}
 			return HttpResponse(t.render(c,request))
 
-# 注销
+# 注销,DONE
 def logout(request):
 	if request.method == 'POST' or request.method == 'GET':
 		request.session.flush()
 		form = LoginForm()
 		return render(request,'login.html',{'form':form,'message':'已注销，请重新登录'})
 
-#修改密码
+#修改密码,DONE
 def resetByUsernameForm(request):
 	if request.method =='GET':
 		request.session.flush()
@@ -123,13 +122,6 @@ def resetByUsernameForm(request):
 
 
 
-
-
-
-
-
-
-
 @ensure_csrf_cookie
 @csrf_exempt
 def main(request):
@@ -148,26 +140,40 @@ def ajax(request):
 		# 	return delete_file(request)
 
 
+
+
 @ensure_csrf_cookie
 @csrf_exempt
 def delete_file(request):
 	print(request.POST)
+	c = {}
 	if request.method =='POST':
+		if not request.POST.get('username'):
+			c['message'] = '请登录'
 		if request.POST.get('file_guid') != None:
-			print(request.POST.get('file_guid'))
+			# print(request.POST.get('file_guid'))
+			# print(request.POST.get('file_guid'))
+			login_user = User.objects.filter(username=request.session.get('username'))[0]	
 			delete_file_guid = request.POST.get('file_guid')
-			delete_file_path = File.objects.filter(file_guid = delete_file_guid)[0].path
-			print('path:'+delete_file_path)
-			if os.path.exists(delete_file_path):  
-   	 			os.remove(delete_file_path)
+			delete_file = File.objects.filter(file_guid = delete_file_guid)[0]
+			if delete_file.owner != login_user:
+				c['message'] = 'あなたはわたしのマスタか？'
+				return JsonResponse(c)
+			delete_file_path = delete_file.path
+			delete_dir_path = '/'.join(delete_file_path.split('/')[:-1])
+			print('path:'+delete_dir_path)
+			if os.path.exists(delete_dir_path):  
+				shutil.rmtree(delete_dir_path)
 			# file表中删除
 			File.objects.filter(file_guid = delete_file_guid).delete()
 			# share表中删除（cascade)///
+			# t = loader.get_template('file_management.html')
+			# files = File.objects.all()
+			# c ={'files':files}
+			c['message'] = 'success'
+	return JsonResponse(c)
 
-			t = loader.get_template('file_management.html')
-			files = File.objects.all()
-			c ={'files':files}
-	return HttpResponse(c)
+
 
 @ensure_csrf_cookie
 @csrf_exempt
@@ -188,7 +194,7 @@ def share_file(request):
 			shared_file = File.objects.filter(file_guid = request.POST.get('file_guid'))[0]
 		owner = User.objects.filter(user_guid = request.session.get('user_guid'))[0]
 		if not owner:
-			return HttpResponse('failed to share')
+			return HttpResponse('请登录')
 		for i in co_editors: # i是输入的字段（用户名或者邮箱）
 			# 只会match到一个，除非A的用户名叫xxx@yyy.com,B的邮箱也是xxx@yyy.com, 在注册时检查一下？
 			if User.objects.filter(Q(username=i) | Q(email=i)).exists():
@@ -200,7 +206,7 @@ def share_file(request):
 				correct_co_editors.append(matched_co_editors.username)
 				share = Share()
 				share.owner = owner
-				share.share_user_guid = matched_co_editors.user_guid
+				share.share_user= matched_co_editors
 				share.shared_file = shared_file
 				share.permission = 'share'
 				share.save()
@@ -221,16 +227,24 @@ def share_file(request):
 
 
 
-
+@csrf_exempt
 def upload(request):
+	c ={}  #return message
 	t = loader.get_template('main.html')
 	if request.method == 'POST':# 获取对象
-		obj = request.FILES.get('myfile')
-		filename = obj.name
-		username = request.session.get('username')
-		print(filename +'\t' + username)
-		c ={}
-		if request.session.get('log_in'):
+		obj = request.FILES.get('file')
+		filename = obj.name	#filename
+		username = request.session.get('username') #username
+		file_dir = '' # full path of the file
+		
+		
+		user = User.objects.filter(username=username)[0] if request.session.get('login') else None
+		# print(filename +'\t' + username)
+		
+		if File.objects.filter(filename=filename).filter(owner=user).count():
+			c['message']='已存在同名文件，请先修改文件名后上传'
+			return JsonResponse(c)
+		if request.session.get('login'):
 			file_dir = BASE_DIR +'/files/'+username+'/'+filename+'/'
 		else:
 			file_dir = BASE_DIR +'/files/temp/'+username+'/'+filename+'/'
@@ -240,17 +254,19 @@ def upload(request):
 			for chunk in obj.chunks():
 				f.write(chunk)
 		if os.path.exists(file_dir+filename):
-			c['messgage']='上传成功'
+			c['message']='success'
 			file =File()
 			file.owner = User.objects.filter(username=username)[0]
 			file.file_guid = str(uuid.uuid1())
 			file.filename = filename
 			file.path = file_dir+filename
+			total_page = file.total_page = get_total_page(file_dir+filename)
 			file.save()
+			convert_next_10(file.path,1,total_page)
 		else:
-			c['messgage']='上传失败'
-
-	return HttpResponse(t.render(c,request))
+			c['message']='上传失败'
+			return JsonResponse(c)
+	return JsonResponse(c)
 
 def get_total_page(file_path):
 	f= PyPDF2.PdfFileReader(file_path)
@@ -258,24 +274,27 @@ def get_total_page(file_path):
 
 
 
-
-
-
-
 def file_management(request):
 	t = loader.get_template('file_management.html')
-	username = request.session.get('username')
-	user = User.objects.filter(username=username)[0]
+	user = User.objects.filter(username=request.session.get('username'))[0]
 	files = File.objects.filter(owner = user)
-	share = Share.objects.filter(share_user_guid = request.session.get('user_guid'))
-	c ={'files':files,'shares':share,}
-	print(files)
+	shares = Share.objects.filter(share_user=user);
+	# print(shares[0].shared_file)
+	# print(files)
+	c ={'messgage':user.username, 'files':files,'shares':shares}
 	return HttpResponse(t.render(c,request))
 
 
+def convert_next_10(file_path,start_page,total_page):
+	for i in range(start_page,min(start_page+10,total_page+1)):
+		# <filename> <page> <zoom> <degree> <output filename> <needle>
+		img_path = '/'.join(file_path.split('/')[:-1]) +'/'+str(i)+'.png'
+		cmd = 'python3 %s %s %s %s %s %s %s' %(BASE_DIR+'/user/static/scripts/convert_to_img.py',file_path, str(i), '400','0',img_path,'1')
+		print(cmd)
+		os.popen(cmd,'w')
+
+def test(request):
+	print(BASE_DIR)
+	return HttpResponse('123')
 
 
-
-
-
-
